Do not complain about 'pkg' being loaded when updating
Of course it is loaded. But this should not cause problems.
As for the pkg dependencies, they currently do not go in
the private library, so they are fine as well. If they are
loaded in the current session, then we get a warning.

After pkg is updated, `pkg_create_private_lib()` can update
the private lib. This will be called automatically if the new
pkg version has new dependencies.

In the future, we could have a dedicated function to update
pkg itself, e.g. when the user does not want pkg dependencies
in their regular library, but for now we can close this.

Closes #52.