Add ccache installation and setup instructions
As a reaction to a question raised in the audience during userR! and the
discussion on #103, an extra section in the README that explains
(hopefully clearly) how to setup ccache for use with R and pak to speed
up installation of packages that need compilation.

Credits to @eddelbuettel for making me aware of this idea. Settings taken
from his blog at

http://dirk.eddelbuettel.com/blog/2017/11/27/

Compared to the post I've reduced maximum cache size from 5G to 1G as
installation of packages tidyverse, devtools and data.table generates
less than 150M of cache files and extended the inline comments
somewhat.