macOS: use pak being built to push packages
Because there is a bug in (deployed) pak that pushes
to the wrong tag, because `push_packages()` assume(d)
that the pak versions doing the push is the same as
the pak version being pushed.

With this change we make sure that this assumption works,
until 83b1cc53c2921fe8716bd25c56ab3c9edd47715d is deployed.