List known platforms in `pkg_sysreqs()` documentation
This is the second time I've had to figure this out in order to generate a working example on my laptop, so I figured I'd actually document it. I couldn't get the documentation to build correctly and I'm not sure what style you use for these documentation helper functions, so you'll need to do a little work to finish it off.