linux-gnu build: use mbedtls instead of openssl
With openssl I ran into issue on Fedora 42, where R links to
libcrypt.so through libtirpc.so. Loading the shared lib of the
curl package crashes here, because somehow the symbols from
the system's dynamic libcrypt.so are used instead of the
statically linked symbols.

With mbedtls the pak build works on Fedora 42 (tried only aarch64
so far).